Add my name and github link to CONTRIBUTION_HISTORY_PRE_OPEN_SOURCE.md
Open sourcing the framework has been journey, years in the making and
I'm overjoyed to see it reach this milestone and finally be able to
share the framework with the world.

We could not have done this without the contributions of everyone
involved and I hope this file serves that purpose well.

I look forward to continuing to develop the framework in the open and
doing my best to assist new contributors in their quest to help improve
the framework.